refactor: 3단계 - 코드리뷰 반영
- 패키지 분리
