+++ src/main/java/ladder/strategy/PointGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.strategy;
 
 @FunctionalInterface
 public interface PointGenerator {
