+++ src/main/java/ladder/type/DirectionType.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.type;
 
 import java.util.function.Function;
